# backend/config/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alice Voice Backend API",
    "description": "REST and socket contract owned by the backend. Frontend copies DTOs from this file.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local daemon"
    }
  ],
  "paths": {
...
    "/api/v1/resolver/gliner/models/{id}": {
      "delete": {
        "tags": [
          "resolver"
        ],
        "summary": "Reply of `DELETE /api/v1/resolver/gliner/models/{id}`.",
        "operationId": "delete_gliner_model_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model id, example: gliner_small-v2.1",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "State of the resolver",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolverStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "The daemon knows no model with that identifier"
          }
        }
      }
    },
    "/api/v1/resolver/gliner/models/{id}/download": {
      "post": {
        "tags": [
          "resolver"
        ],
        "summary": "Reply of `POST /api/v1/resolver/gliner/models/{id}/download`.",
        "description": "The download runs in the background, so the reply carries the state the\nsettings page polls while it runs.",
        "operationId": "download_gliner_model_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model id, example: gliner_small-v2.1",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "State of the resolver",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolverStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "The daemon knows no model with that identifier"
          },
          "409": {
            "description": "A download already runs"
          }
        }
      }
    },
    "/api/v1/resolver/local/models/{id}": {
      "delete": {
        "tags": [
          "resolver"
        ],
        "summary": "Reply of `DELETE /api/v1/resolver/local/models/{id}`.",
        "operationId": "delete_local_model_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model id, example: bge-small-en-v1.5",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "State of the resolver",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolverStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "The daemon knows no model with that identifier"
          }
        }
      }
    },
    "/api/v1/resolver/local/models/{id}/download": {
      "post": {
        "tags": [
          "resolver"
        ],
        "summary": "Reply of `POST /api/v1/resolver/local/models/{id}/download`.",
        "description": "The download runs in the background, so the reply carries the state the\nsettings page polls while it runs.",
        "operationId": "download_local_model_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model id, example: bge-small-en-v1.5",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "State of the resolver",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolverStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "The daemon knows no model with that identifier"
          },
          "409": {
            "description": "A download already runs"
          }
        }
      }
    },
    "/api/v1/resolver/preview": {
      "post": {
        "tags": [
          "resolver"
        ],
        "summary": "Reply of `POST /api/v1/resolver/preview`.",
        "description": "The preview reads one message with the stored settings and runs\nnothing, so the settings page can show the route of a sentence before\nthe user sends it. The reply carries the metadata of a stored turn, so\nthe route of a sentence and the route of a message the daemon really\nhandled read the same way.",
        "operationId": "preview_message_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolverPreviewRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "How the resolver would read the message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolverPreviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "413": {
            "description": "Message too long"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlinerDownloadDto": {
        "type": "object",
        "description": "The GLiNER download that runs right now.",
        "required": [
          "model",
          "receivedBytes",
          "done"
        ],
        "properties": {
          "done": {
            "type": "boolean",
            "description": "Whether the download finished."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "The error that stopped the download, or null."
          },
          "model": {
            "type": "string",
            "description": "Identifier of the model."
          },
          "receivedBytes": {
            "type": "integer",
            "format": "int64",
            "description": "Bytes written so far.",
            "minimum": 0
          },
          "totalBytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Size the server announced, or null when it announced none.",
            "minimum": 0
          }
        }
      },
      "GlinerModelDto": {
        "type": "object",
        "description": "One GLiNER model the daemon can download.",
        "required": [
          "id",
          "name",
          "note",
          "sizeBytes",
          "installed"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier of the model, for example `gliner_small-v2.1`."
          },
          "installed": {
            "type": "boolean",
            "description": "Whether the model is on disk."
          },
          "name": {
            "type": "string",
            "description": "Name the settings page shows."
          },
          "note": {
            "type": "string",
            "description": "One sentence about the model."
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64",
            "description": "Size of the download in bytes.",
            "minimum": 0
          }
        }
      },
      "GlinerStoreDto": {
        "type": "object",
        "description": "State of the built in GLiNER models.",
        "required": [
          "model",
          "device",
          "activeDevice",
          "installed",
          "threshold",
          "modelsDir",
          "devices",
          "cudaBuild",
          "models"
        ],
        "properties": {
          "activeDevice": {
            "type": "string",
            "description": "The device the selected model would run on now."
          },
          "cudaBuild": {
            "type": "boolean",
            "description": "Whether this build carries CUDA support."
          },
          "device": {
            "type": "string",
            "description": "The device the selected model should run on."
          },
          "devices": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Devices this build and this machine can run on, best first."
          },
          "download": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GlinerDownloadDto",
                "description": "The download that runs right now, or null."
              }
            ]
          },
          "installed": {
            "type": "boolean",
            "description": "Whether the selected model is installed."
          },
          "model": {
            "type": "string",
            "description": "Identifier of the selected model."
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GlinerModelDto"
            },
            "description": "Every model the daemon can download."
          },
          "modelsDir": {
            "type": "string",
            "description": "Directory that holds the downloaded models."
          },
          "threshold": {
            "type": "number",
            "format": "float",
            "description": "Smallest probability a label needs to count."
          }
        }
      },
...
      "LabelBudgetDto": {
        "type": "object",
        "description": "How many labels the intent configuration adds up to.",
        "required": [
          "labels",
          "softLimit",
          "hardLimit",
          "intents",
          "overSoft",
          "overHard"
        ],
        "properties": {
          "hardLimit": {
            "type": "integer",
            "description": "Number of labels a GLiNER model reads before it degrades.",
            "minimum": 0
          },
          "intentLimit": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Number of intents the resolver can offer, or null when it has no\nlimit. The layered router reads a catalog of any size, because\nonly the short list of one turn reaches a model.",
            "minimum": 0
          },
          "intents": {
            "type": "integer",
            "description": "Number of configured intents.",
            "minimum": 0
          },
          "labels": {
            "type": "integer",
            "description": "Number of labels a GLiNER model reads for this configuration.\nOne label per intent, plus one per entity, plus one per value of\na closed entity.",
            "minimum": 0
          },
          "overHard": {
            "type": "integer",
            "description": "Number of labels that are past the hard limit, or zero.",
            "minimum": 0
          },
          "overSoft": {
            "type": "integer",
            "description": "Number of labels that are past the soft limit, or zero.",
            "minimum": 0
          },
          "softLimit": {
            "type": "integer",
            "description": "Number of labels a GLiNER model reads comfortably.",
            "minimum": 0
          },
          "warning": {
            "type": [
              "string",
              "null"
            ],
            "description": "The warning the settings page shows, or null."
          }
        }
      },
...
      "LlamaStatusDto": {
        "type": "object",
        "description": "Configuration of the llama.cpp resolver.",
        "required": [
          "baseUrl",
          "model",
          "intentLimit"
        ],
        "properties": {
          "baseUrl": {
            "type": "string",
            "description": "Base URL of the llama.cpp server."
          },
          "intentLimit": {
            "type": "integer",
            "description": "Number of intents the resolver can offer.",
            "minimum": 0
          },
          "model": {
            "type": "string",
            "description": "Model name the server answers to."
          }
        }
      },
...
      "ResolverPreviewDto": {
        "type": "object",
        "description": "Reply of `POST /api/v1/resolver/preview`.\n\nThe preview carries the same metadata a stored turn carries, so a\nsentence the user tries is read with the same report as a message the\ndaemon really handled, minus the run. The intent it met, the reader\nthat read every value, and the route the message took are the report a\nuser tunes the pipeline by.",
        "required": [
          "text",
          "matched",
          "reply"
        ],
        "properties": {
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "The probability of the choice, or null when the resolver sent none."
          },
          "intent": {
            "type": [
              "string",
              "null"
            ],
            "description": "The name of the intent the resolver met, or null when it met none."
          },
          "matched": {
            "type": "boolean",
            "description": "True when the resolver met an intent."
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MessageMetaDto",
                "description": "How the daemon read the turn, or null when it read nothing at all."
              }
            ]
          },
          "reply": {
            "type": "string",
            "description": "The reply the daemon would store, or empty when the turn would run\nthe command of its intent."
          },
          "text": {
            "type": "string",
            "description": "The message the daemon read."
          }
        }
      },
      "ResolverPreviewRequestDto": {
        "type": "object",
        "description": "Body of `POST /api/v1/resolver/preview`.\n\nA preview reads one message the way a turn would and runs nothing, so\nthe settings page can show the route of a sentence before the user\nsends it.",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "description": "The message to read."
          }
        }
      },
      "ResolverStatusDto": {
        "type": "object",
        "description": "Reply of `GET /api/v1/resolver`.\nThis tells the settings page which engines are ready to work.",
        "required": [
          "backend",
          "llama",
          "gliner",
          "budget",
          "router"
        ],
        "properties": {
          "backend": {
            "type": "string",
            "description": "The engine that reads the intent of a message."
          },
          "budget": {
            "$ref": "#/components/schemas/LabelBudgetDto",
            "description": "How many labels the intent configuration adds up to."
          },
          "gliner": {
            "$ref": "#/components/schemas/GlinerStoreDto",
            "description": "State of the built in GLiNER models."
          },
          "llama": {
            "$ref": "#/components/schemas/LlamaStatusDto",
            "description": "Configuration of the llama.cpp resolver."
          },
          "router": {
            "$ref": "#/components/schemas/RouterStatusDto",
            "description": "The state of the layered router."
          }
        }
      },
...
      "RouterStatusDto": {
        "type": "object",
        "description": "The state of the layered router.\n\nThe router has a stage per setting, and each stage names the reader it\nwould use right now. The settings page shows that state, so a stage\nthat cannot run its reader says so before a turn depends on it.",
        "required": [
          "fastPath",
          "retrieve",
          "decide",
          "extract",
          "topK",
          "floor",
          "margin",
          "lexicalWeight",
          "denseWeight",
          "embedModel",
          "embedSource",
          "embedLocalModel",
          "rerankModel",
          "layaModel",
          "localDevice",
          "activeLocalDevice",
          "local",
          "phraseGate",
          "listMatch",
          "listFloor",
          "embeddingsReachable"
        ],
        "properties": {
          "activeLocalDevice": {
            "type": "string",
            "description": "The device a built in model of the router would run on now."
          },
          "decide": {
            "type": "string",
            "description": "How the decision pass chooses one of the short list."
          },
          "denseWeight": {
            "type": "number",
            "format": "float",
            "description": "Weight of the embeddings of the catalog."
          },
          "embedLocalModel": {
            "type": "string",
            "description": "Identifier of the built in embedding model."
          },
          "embedModel": {
            "type": "string",
            "description": "Model name the embedding server answers to."
          },
          "embedSource": {
            "type": "string",
            "description": "Where the retrieval stage and the list matcher read the vectors."
          },
          "embeddingsDetail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the model server answers with no embeddings, or null."
          },
          "embeddingsReachable": {
            "type": "boolean",
            "description": "Whether the model server answers the router with embeddings."
          },
          "extract": {
            "type": "string",
            "description": "How the values of the entities of the chosen intent are read."
          },
          "fastPath": {
            "type": "boolean",
            "description": "Whether the deterministic pass reads the message first."
          },
          "floor": {
            "type": "number",
            "format": "float",
            "description": "Smallest score the decision pass accepts."
          },
          "layaModel": {
            "type": "string",
            "description": "Identifier of the built in decision model."
          },
          "lexicalWeight": {
            "type": "number",
            "format": "float",
            "description": "Weight of the words of the catalog."
          },
          "listFloor": {
            "type": "number",
            "format": "float",
            "description": "Smallest cosine similarity an embedding match of a value needs."
          },
          "listMatch": {
            "type": "string",
            "description": "How a mention is read against the values of an entity."
          },
          "local": {
            "$ref": "#/components/schemas/LocalStoreDto",
            "description": "State of the built in models of the router."
          },
          "localDevice": {
            "type": "string",
            "description": "The device a built in model of the router runs on."
          },
          "margin": {
            "type": "number",
            "format": "float",
            "description": "Smallest distance between the best and the second best."
          },
          "phraseGate": {
            "type": "boolean",
            "description": "Whether a phrase counts only when the message shares its action."
          },
          "rerankModel": {
            "type": "string",
            "description": "Identifier of the built in reranker."
          },
          "retrieve": {
            "type": "string",
            "description": "The evidence the retrieval pass ranks the catalog with."
          },
          "topK": {
            "type": "integer",
            "description": "Size of the short list the retrieval pass keeps.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}